Update gitignore from templates
The master .gitignore was originally generated from GitHub's
templates at https://github.com/github/gitignore but had fallen
out of date. This syncs up with the changes there, and adds in
the Maven.gitignore template as well.

The official Java template excluded all packaged files,
including .jar and .zip which the repo does contain (e.g.
the SWT packages in `uis/lib/`), so I commented those out
with a note.

Had no effect on the `git status` output in my local clone, but
that's as it should be really.